show coproces stdout as background process

+++ load.sh
@@ -30,26 +30,36 @@ EOF
 
     sed -r -i -e "s/\\\$RecoverySecret/$RecoverySecret/" /docker-entrypoint-initdb.d/90-create_test_user.sql
 
-    coproc tailcop { exec docker-entrypoint.sh $MYSQLOPTS --skip-networking 2>&1 ; }
+    coproc tailcop {
+        exec docker-entrypoint.sh $MYSQLOPTS --skip-networking 2>&1 
+    }
 
-    sleep 10800 && echo "$(date '+%m/%d %H:%M:%S'): Timeout during init" && kill $tailcop_PID &
+    coproc timeout {
+        sleep 7200 &&
+        echo "$(date '+%m/%d %H:%M:%S'): Timeout during init" >&4 &&
+        kill $tailcop_PID 
+    } 4>&2
 
-    while read -ru ${tailcop[0]} line; do
+    exec 3<&${tailcop[0]}
+
+    while read -ru 3 line; do
         echo $line
         [ $(expr "$line" : 'MySQL init process done. Ready for start up.') -gt 0 ] && break
     done
 
-    while read -ru ${tailcop[0]} line; do
+    while read -ru 3 line; do
         echo $line
         [ $(expr "$line" : '.*\[Note\] mysqld: ready for connections.') -gt 0 ] && break
     done
-    sleep 1
+
+    # Keep reading and showing myqld messages
+    cat <&3 &
+
     # Init completed, kill the timeout killer
-    pkill -x sleep
+    [ -n "$timeout_PID" ] && kill $timeout_PID
+
     echo "$(date '+%m/%d %H:%M:%S'): Shutting down MySQL Server"
-    kill $tailcop_PID
-    # Wait for shutdown while showing progress
-    cat <&${tailcop[0]}
+    [ -n "$tailcop_PID" ] && kill $tailcop_PID && wait $tailcop_PID
 else
     # Delegate control to docker-entrypoint.sh
     exec /usr/local/bin/docker-entrypoint.sh $MYSQLOPTS
